+++ tests/test_dataset.py
@@ -26,7 +26,7 @@ TEST_CASE_1 = [(128, 128, 128)]
 class TestDataset(unittest.TestCase):
     @parameterized.expand([TEST_CASE_1])
     def test_shape(self, expected_shape):
-        test_image = nib.Nifti1Image(np.random.randint(0, 2, size=[128, 128, 128]), np.eye(4))
+        test_image = nib.Nifti1Image(np.random.randint(0, 2, size=[128, 128, 128]).astype(float), np.eye(4))
         with tempfile.TemporaryDirectory() as tempdir:
             nib.save(test_image, os.path.join(tempdir, "test_image1.nii.gz"))
             nib.save(test_image, os.path.join(tempdir, "test_label1.nii.gz"))
